+++ crates/models/deploy/legacy_logs.rs
@@ -106,6 +106,7 @@ pub struct DepositReport {
     pub token_id: u32,
     pub time: u32,
     pub amount: i64,
+    pub custom_id: i64,
 }
 
 #[repr(C)]
@@ -179,6 +180,7 @@ pub struct WithdrawReport {
     pub token_id: u32,
     pub time: u32,
     pub amount: i64,
+    pub custom_id: i64,
 }
 
 #[repr(C)]