+++ crates/models/src/instruction_constants.rs
@@ -367,20 +367,26 @@ impl DrvInstruction for ExtendCandles {
     const MIN_ACCOUNTS: usize = 4;
 }
 
+pub struct WithdrawSwapFeesInstruction;
+impl DrvInstruction for WithdrawSwapFeesInstruction {
+    const INSTRUCTION_NUMBER: u8 = 74;
+    const MIN_ACCOUNTS: usize = 12;
+}
+
 pub struct SetSMAMinQtyInstruction;
 impl DrvInstruction for SetSMAMinQtyInstruction {
-    const INSTRUCTION_NUMBER: u8 = 74;
+    const INSTRUCTION_NUMBER: u8 = 75;
     const MIN_ACCOUNTS: usize = 3;
 }
 
 pub struct ChangeSAMFeesPolictyInstruction;
 impl DrvInstruction for ChangeSAMFeesPolictyInstruction {
-    const INSTRUCTION_NUMBER: u8 = 75;
+    const INSTRUCTION_NUMBER: u8 = 76;
     const MIN_ACCOUNTS: usize = 3;
 }
 
 pub struct SuspendInstrumentInstruction;
 impl DrvInstruction for SuspendInstrumentInstruction {
-    const INSTRUCTION_NUMBER: u8 = 76;
+    const INSTRUCTION_NUMBER: u8 = 77;
     const MIN_ACCOUNTS: usize = 4;
 }